backend/requirements.txt
fastapi==0.91.0
uvicorn==0.20.0
pymongo==3.11.0
python-dotenv==1.0.1
dnspython==1.16.0
